requirements.txt
requests==2.22.0
parsel==1.6.0
typer==0.3.2
pyyaml==5.3.1
click-spinner==0.1.10
tabulate==0.8.7
beautifulsoup4==4.9.3
pika==1.2.0
selenium==3.141.0
simple-term-menu==0.10.5
psutil==5.8.0
webdriver-manager==3.3.0
kubernetes==12.0.1